Change SharedPreferencesHelper to use FoodItemEntity
This was necessary because Gson couldn't create instances of FoodItemEntity
when deserialising a list the items of which were serialised as FoodItem. In
other words, Gson cannot instantiate interfaces.

As a consequence, list types in GroceryListState had to be changed also, but
parameters of food-item related functions could remain thanks to casting.