Fix countdown values of items not incrementing
This was due to the previous incrementation mechanism triggering the
incrementing when MainActivity was created on non-grocery days.

Now that we're on a single activity architecture, MainActivity never
gets destroyed and neither does GroceryListFragment, so the mechanism
stops functioning.

The solution was to explicitly trigger the incrementing when
GroceryListFragment is paused.